+++ src/main/java/com/chinamobile/iot/lightapp/mysql/model/CheckItemScoreExample.java
@@ -394,52 +394,52 @@ public class CheckItemScoreExample {
             return (Criteria) this;
         }
 
-        public Criteria andScoreEqualTo(Float value) {
+        public Criteria andScoreEqualTo(Integer value) {
             addCriterion("score =", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreNotEqualTo(Float value) {
+        public Criteria andScoreNotEqualTo(Integer value) {
             addCriterion("score <>", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreGreaterThan(Float value) {
+        public Criteria andScoreGreaterThan(Integer value) {
             addCriterion("score >", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreGreaterThanOrEqualTo(Float value) {
+        public Criteria andScoreGreaterThanOrEqualTo(Integer value) {
             addCriterion("score >=", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreLessThan(Float value) {
+        public Criteria andScoreLessThan(Integer value) {
             addCriterion("score <", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreLessThanOrEqualTo(Float value) {
+        public Criteria andScoreLessThanOrEqualTo(Integer value) {
             addCriterion("score <=", value, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreIn(List<Float> values) {
+        public Criteria andScoreIn(List<Integer> values) {
             addCriterion("score in", values, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreNotIn(List<Float> values) {
+        public Criteria andScoreNotIn(List<Integer> values) {
             addCriterion("score not in", values, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreBetween(Float value1, Float value2) {
+        public Criteria andScoreBetween(Integer value1, Integer value2) {
             addCriterion("score between", value1, value2, "score");
             return (Criteria) this;
         }
 
-        public Criteria andScoreNotBetween(Float value1, Float value2) {
+        public Criteria andScoreNotBetween(Integer value1, Integer value2) {
             addCriterion("score not between", value1, value2, "score");
             return (Criteria) this;
         }
